// src/app/shared/generated/model/createSearchConfigRequest.ts
/**
 * onecx-concessions bff
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateSearchConfigRequest { 
    page: string;
    name: string;
    /**
     * Version increment of the fields in the UI which you should use when you are making incompatible changes to those fields.
     */
    fieldListVersion: number;
    /**
     * Defines whether this config can be changed in the UI
     */
    isReadonly: boolean;
    /**
     * Indicates whether the advanced mode should be displayed
     */
    isAdvanced: boolean;
    columns: Array<string>;
    values: { [key: string]: string; };
}